+++ tests/ReindexTest.php
@@ -151,24 +151,6 @@ class ReindexTest extends Base
 
     /**
      * @group functional
-     * @group legacy
-     */
-    public function testReindexWithDeprecatedConstantSetOnWaitForCompletion(): void
-    {
-        $oldIndex = $this->_createIndex('idx1', true, 2);
-        $this->_addDocs($oldIndex, 10);
-
-        $newIndex = $this->_createIndex('idx2', true, 2);
-
-        $reindex = new Reindex($oldIndex, $newIndex);
-        $reindex->setWaitForCompletion(Reindex::WAIT_FOR_COMPLETION_FALSE);
-        $reindex->run();
-
-        $this->assertNotEmpty($reindex->getTaskId());
-    }
-
-    /**
-     * @group functional
      */
     public function testReindexWithScript(): void
     {